sftp bug fixes

+++ step1/download_from_ftp.py
@@ -16,6 +16,33 @@ from django.contrib.auth.decorators import login_required
 from django.views.decorators.csrf import csrf_exempt
 
 
+# This function will download each file/dir from SFTP server to the local directory
+# in case of given folder is blank or have any kind of error this will return False
+def download_directory_from_ftp(ftp_connection, remote_directory, local_directory):
+    if not os.path.exists(local_directory):
+        os.makedirs(local_directory)
+
+    content_list = ftp_connection.nlst(remote_directory)
+
+    if len(content_list) == 0:
+        return False
+    
+    try:
+        for entry in content_list:
+            remote_path = os.path.join(remote_directory, entry)
+            local_path = os.path.join(local_directory, entry)
+            
+            if is_ftp_content_folder(ftp_connection, remote_path):
+                download_directory_from_ftp(ftp_connection, remote_path, local_path)
+            else:
+                with open(local_path, 'wb') as local_file:
+                    ftp_connection.retrbinary(f'RETR {entry}', local_file.write)
+
+        return True
+    except Exception as e:
+        print(e)
+        return False
+
 # function to download file
 def download_file(ftp_connection, article, item):
     file_size = ftp_connection.size(article)
@@ -53,50 +80,53 @@ def download_file(ftp_connection, article, item):
     
 
 # function to download folder with its content and convert to zip, finally save to table
-def download_folder_from_ftp_and_save_zip(article, item):
-    article = article + '.zip'
-    zip_name = item.provider.official_name
-    # Create a temporary directory to store downloaded files
+def download_folder_from_ftp_and_save_zip(ftp_connection, article, item):
     temp_dir = 'temp_download/' + item.provider.official_name
-    if not os.path.exists(temp_dir):
-        os.makedirs(temp_dir)
+    state = download_directory_from_ftp(ftp_connection, article, temp_dir)
 
-    # Create a zip file from the downloaded content
-    zip_filename = os.path.join(temp_dir, zip_name)
-    zipped_file = shutil.make_archive(zip_filename.split('.')[0], 'zip', temp_dir)
+    if state:
+        article = article + '.zip'
+        zip_name = item.provider.official_name
+        # Create a temporary directory to store downloaded files
+        if not os.path.exists(temp_dir):
+            os.makedirs(temp_dir)
 
-    x = Archive.objects.filter(file_name_on_source=article)
+        # Create a zip file from the downloaded content
+        zip_filename = os.path.join(temp_dir, zip_name)
+        zipped_file = shutil.make_archive(zip_filename.split('.')[0], 'zip', temp_dir)
 
-    # if file not exists in database, create new record
-    if not(x.exists()):
-        # Save the zip file to the Django model
-        with open(zipped_file, 'rb') as zip_file:
-            zip_file.seek(0)
-            # zip_file_model = ZipFileModel(name=zip_name)
-            # zip_file_model.zip_file.save(zip_name + '.zip', ContentFile(zip_file.read()), save=True)
+        x = Archive.objects.filter(file_name_on_source=article)
 
-            x = Archive.objects.create(
-                file_name_on_source = article,
-                provider = item.provider,
-                processed_on = datetime.datetime.now(tz=pytz.utc),
-                status = 'active',
-                file_size = os.path.getsize(zipped_file),
-                file_type = '.zip'
-            )
+        # if file not exists in database, create new record
+        if not(x.exists()):
+            # Save the zip file to the Django model
+            with open(zipped_file, 'rb') as zip_file:
+                zip_file.seek(0)
+                # zip_file_model = ZipFileModel(name=zip_name)
+                # zip_file_model.zip_file.save(zip_name + '.zip', ContentFile(zip_file.read()), save=True)
 
-            article = str(x.id) + '.' + article.split('.')[-1]
-            x.file_content.save(article, zip_file)
+                x = Archive.objects.create(
+                    file_name_on_source = article,
+                    provider = item.provider,
+                    processed_on = datetime.datetime.now(tz=pytz.utc),
+                    status = 'active',
+                    file_size = os.path.getsize(zipped_file),
+                    file_type = '.zip'
+                )
 
-        # Cleanup temporary directory and return
-        shutil.rmtree(temp_dir)
-        return
+                article = str(x.id) + '.' + article.split('.')[-1]
+                x.file_content.save(article, zip_file)
 
-    # if file exists than check the file size. If file size is different update the existing record
-    if (x.exists() and not (x[0].file_size == os.path.getsize(zipped_file))):
-        x[0].status = 'active'
-        with open(zipped_file, 'rb') as zip_file:
-            zip_file.seek(0)
-            x[0].file_content.save(article, zip_file)
+            # Cleanup temporary directory and return
+            shutil.rmtree(temp_dir)
+            return
+
+        # if file exists than check the file size. If file size is different update the existing record
+        if (x.exists() and not (x[0].file_size == os.path.getsize(zipped_file))):
+            x[0].status = 'active'
+            with open(zipped_file, 'rb') as zip_file:
+                zip_file.seek(0)
+                x[0].file_content.save(article, zip_file)
             
     # Cleanup temporary directory
     shutil.rmtree(temp_dir)
@@ -113,7 +143,7 @@ def download_from_ftp(request):
     # get all providers that are due to be accessed
     due_for_download = Provider_meta_data_FTP.objects.filter(
         provider__next_due_date__lte = datetime.datetime.now(tz=pytz.utc)
-        )
+        ).exclude(protocol='SFTP')
     # .exclude(provider__working_name="CSIRO")
     
     # if none is due to be accessed abort the process
@@ -128,7 +158,6 @@ def download_from_ftp(request):
 
     # if providers are due to be accessed
     for item in due_for_download:
-        print(item.provider.working_name)
         # try to ftp_connection to FTP, if error occures update the status to Provider_meta_data_FTP and continue to access next FTP
         try:
             ftp_connection = ftplib.FTP(item.server)
@@ -153,7 +182,7 @@ def download_from_ftp(request):
                     # check if the article is file or directory
                     if is_ftp_content_folder(ftp_connection, article):
                         # download the folder
-                        download_folder_from_ftp_and_save_zip(article, item)
+                        download_folder_from_ftp_and_save_zip(ftp_connection, article, item)
                     else:
                         # download the file
                         download_file(ftp_connection, article, item)
